pyproject.toml
[tool.poetry]
name = "llm-distillery"
version = "0.1.0"
description = "Transform large language model expertise into fast, specialized local models"
authors = ["Your Name <your.email@example.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
torch = "^2.1.0"
transformers = "^4.35.0"
datasets = "^2.14.0"
anthropic = "^0.8.0"
google-generativeai = "^0.3.0"
openai = "^1.6.0"
pydantic = "^2.5.0"
tqdm = "^4.66.0"
pandas = "^2.1.0"
numpy = "^1.26.0"
scikit-learn = "^1.3.0"
wandb = "^0.16.0"
pyyaml = "^6.0.1"
jsonlines = "^4.0.0"
fastapi = "^0.104.0"
uvicorn = "^0.24.0"
python-multipart = "^0.0.6"
safetensors = "^0.4.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.0"
pytest-cov = "^4.1.0"
black = "^23.11.0"
ruff = "^0.1.6"
mypy = "^1.7.0"
ipython = "^8.17.0"
jupyter = "^1.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 100
target-version = ['py310']

[tool.ruff]
line-length = 100
target-version = "py310"
select = ["E", "F", "W", "I", "N", "UP"]

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
python_classes = "Test*"
python_functions = "test_*"
